% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ballot_question.R
\docType{data}
\name{ballot_question}
\alias{ballot_question}
\title{California Gubernatorial Recall Election - Statement of Vote, September 14, 2021.}
\format{
\subsection{\code{ballot_measures}}{

A data frame with 59 entries and 5 columns:
\describe{
\item{County}{County in CA}
\item{Yes_num}{Number of 'yes' votes for a county for the recall ballot measures}
\item{No_num}{Number of 'no' votes for a county for the recall ballot measures}
\item{Yes_per}{Percentage of 'yes' votes for a county for the recall ballot measures}
\item{No_per}{Percentage of 'no' votes for a county for the recall ballot measures}
...
}
}
}
\source{
\url{https://www.sos.ca.gov/elections/prior-elections/statewide-election-results/2021-ca-gov-recall/statement-vote}
}
\usage{
ballot_question
}
\description{
A tidy dataset of the subset of data for the results of the ballot measures voted on during the 2021 California Gubernatorial Recall Election.
Data were tidied using Microsoft Excel.
}
\keyword{datasets}
