% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_summary.R
\name{quick_summary_vaxx_vs_recall}
\alias{quick_summary_vaxx_vs_recall}
\title{Quickly summarize how a county of your choice compares to most extreme California counties in terms of vaccination rates and support for the 2021 recall of Gov. Gavin Newsom.}
\usage{
quick_summary_vaxx_vs_recall(yourcounty)
}
\arguments{
\item{yourcounty}{A string.}
}
\value{
A string.
}
\description{
Quickly summarize how a county of your choice compares to most extreme California counties in terms of vaccination rates and support for the 2021 recall of Gov. Gavin Newsom.
}
\examples{
quick_summary_vaxx_vs_recall("Los Angeles")
quick_summary_vaxx_vs_recall("Orange")
quick_summary_vaxx_vs_recall("San Mateo")
}
