% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ballot_question_RAW.R
\docType{data}
\name{ballot_measures}
\alias{ballot_measures}
\title{California Gubernatorial Recall Election - Statement of Vote, September 14, 2021.}
\format{
\subsection{\code{ballot_measures}}{

A data frame with 236 entries and 3 columns:
\describe{
\item{Column 1}{County number of votes and county percentage of votes}
\item{Column 2}{Number and percentage of 'yes' votes}
\item{Column 3}{Number and percentage of 'no' votes}
...
}
}
}
\source{
\url{https://www.sos.ca.gov/elections/prior-elections/statewide-election-results/2021-ca-gov-recall/statement-vote}
}
\usage{
ballot_measures
}
\description{
A subset of data for the results of the ballot measures voted on during the 2021 California Gubernatorial Recall Election.
}
\keyword{datasets}
